+++ src/main/java/com/maxmind/minfraud/request/TransactionReport.java
@@ -41,23 +41,42 @@ public final class TransactionReport extends AbstractModel {
         String transactionId;
 
         /**
-         * The constructor for the {@code TransactionReport.Builder} class
+         * The constructor for the {@code TransactionReport.Builder} class with IP address.
          *
          * @param ipAddress The IP address associated with the device used by the customer in the
          *                  transaction.
          * @param tag       A string indicating the likelihood that a transaction may be
          *                  fraudulent.
+         * @deprecated      Use {@link #Builder(Tag)} instead and set the IP address using
+         *                  {@link #ipAddress(InetAddress)}.
          */
+        @Deprecated
         public Builder(InetAddress ipAddress, Tag tag) {
-            if (ipAddress == null) {
-                throw new IllegalArgumentException("ipAddress must not be null");
-            }
+            this(tag);
+            this.ipAddress = ipAddress;
+        }
+
+        /**
+         * The constructor for the {@code TransactionReport.Builder} class.
+         *
+         * @param tag A string indicating the likelihood that a transaction may be fraudulent.
+         */
+        public Builder(Tag tag) {
             if (tag == null) {
                 throw new IllegalArgumentException("tag must not be null");
             }
 
-            this.ipAddress = ipAddress;
             this.tag = tag;
+        }
+
+        /**
+         * @param ipAddress The IP address associated with the device used by the customer in the
+         *                  transaction.
+         * @return The builder object.
+         */
+        public TransactionReport.Builder ipAddress(InetAddress ipAddress) {
+            this.ipAddress = ipAddress;
+            return this;
         }
 
         /**
@@ -73,8 +92,9 @@ public final class TransactionReport extends AbstractModel {
         /**
          * @param maxmindId A unique eight character string identifying a minFraud Standard or
          *                  Premium request. These IDs are returned in the maxmindID field of a
-         *                  response for a successful minFraud request. This field is not required,
-         *                  but you are encouraged to provide it, if possible.
+         *                  response for a successful minFraud request. This field is not required
+         *                  if you provide at least one of the transaction's ipAddress, minfraudId,
+         *                  or transactionId. You are encouraged to provide it, if possible.
          * @return The builder object.
          */
         public TransactionReport.Builder maxmindId(String maxmindId) {
@@ -87,11 +107,12 @@ public final class TransactionReport extends AbstractModel {
         }
 
         /**
-         * @param minfraudId A UUID that identifies a minFraud Score, minFraud Insights, or minFraud
-         *                   Factors request. This ID is returned via getId() in the Score, Insights
-         *                   or Factors response object. This field is not required, but you are
-         *                   encouraged to provide it if the request was made to one of these
-         *                   services.
+         * @param minfraudId A UUID that identifies a minFraud Score, minFraud Insights, or
+         *                   minFraud Factors request. This ID is returned via getId() in the
+         *                   Score, Insights or Factors response object. This field is not
+         *                   required if you provide at least one of the transaction's ipAddress,
+         *                   maxmindId, or minfraudId. You are encouraged to provide it if the
+         *                   request was made to one of these services.
          * @return The builder object.
          */
         public TransactionReport.Builder minfraudId(UUID minfraudId) {
@@ -109,7 +130,9 @@ public final class TransactionReport extends AbstractModel {
         }
 
         /**
-         * @param transactionId The transaction ID you originally passed to minFraud.
+         * @param transactionId The transaction ID you originally passed to minFraud. This field is
+         *                      not required if you provide at least one of the transaction's
+         *                      ipAddress, maxmindId, or minfraudId.
          * @return The builder object.
          */
         public TransactionReport.Builder transactionId(String transactionId) {
@@ -120,8 +143,20 @@ public final class TransactionReport extends AbstractModel {
         /**
          * @return An instance of {@code TransactionReport} created from the fields set on this
          *     builder.
+         * @throws IllegalStateException if none of the required fields are set.
          */
         public TransactionReport build() {
+            if (this.ipAddress == null
+                && this.minfraudId == null
+                && (this.maxmindId == null || this.maxmindId.isEmpty())
+                && (this.transactionId == null || this.transactionId.isEmpty())
+            ) {
+                throw new IllegalArgumentException(
+                    "You must pass at least one of the following: "
+                    + "ipAddress, minfraudId, maxmindId, transactionId."
+                );
+            }
+
             return new TransactionReport(this);
         }
     }